Create .gitignore
(.gitignore is a file that serves to avoid tracking files that do not need to be tracked (for example, assets used in the game) or local settings that either already can't or shouldn't affect other users. Don't be shy about removing this if it DOES cause issues or just if you don't want it here, since:)

Initial commit to confirm that this works!