Added non-functional code for Upgrades.
This commit adds code which adds a log message to the singleton objects, a DecreaseScore method in GameManager, a ScoreDisplay,turns the Circle into a singleton class as well as adds a method to upgrade it, and a skeleton for the Upgrade object which is currently non-functional. Capped off with a save to SampleScene. Good night!